fix page and showcase quick delete buttons so they dont delete those in the site path

+++ app/services/destroy/page.rb
@@ -2,15 +2,19 @@ module Destroy
   class Page
     # Destroy the object only
     def destroy!(page:)
-      page.destroy!
+      if PageQuery.can_delete_page?(page)
+        page.destroy!
+      end
     end
 
     # There are no additional cascades for Pages,
     # so destroys the object only
     def cascade!(page:)
-      ActiveRecord::Base.transaction do
-        DestroyPageItemAssociations.call(page_id: page.id)
-        page.destroy!
+      if PageQuery.can_delete_page?(page)
+        ActiveRecord::Base.transaction do
+          DestroyPageItemAssociations.call(page_id: page.id)
+          page.destroy!
+        end
       end
     end
   end
